fix(firmware): componente json do ESP-IDF no lugar de cJSON
- cJSON é fornecido pelo componente 'json' no ESP-IDF
- adiciona esp_timer aos PRIV_REQUIRES e cast explícito de strlen

+++ Esp32/main/main.c
@@ -105,7 +105,7 @@ static void post_reading(float temperature_c)
 
     esp_http_client_handle_t client = esp_http_client_init(&config);
     esp_http_client_set_header(client, "Content-Type", "application/json");
-    esp_http_client_set_post_field(client, payload, strlen(payload));
+    esp_http_client_set_post_field(client, payload, (int)strlen(payload));
 
     esp_err_t err = esp_http_client_perform(client);
     if (err == ESP_OK) {
